chore(nonblocking/coroutines): change package name
