+++ unRealEstate-NonBlocking-Coroutines/src/main/kotlin/dev/hosi/unrealestatenonblockingcoroutines/UnRealEstateNonBlockingCoroutinesApplication.kt
@@ -1,0 +1,11 @@
+package dev.hosi.unrealestatenonblockingcoroutines
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class UnRealEstateNonBlockingApplication
+
+fun main(args: Array<String>) {
+    runApplication<UnRealEstateNonBlockingApplication>(*args)
+}
