+++ unRealEstate-NonBlocking-Coroutines/src/main/kotlin/dev/hosi/unrealestatenonblockingcoroutines/dtos/AgentDTO.kt
@@ -1,0 +1,28 @@
+package dev.hosi.unrealestatenonblockingcoroutines.dtos
+
+import dev.hosi.unrealestatenonblockingcoroutines.entities.Agent
+import java.time.LocalDate
+
+data class AgentDTO(
+    val id: Long,
+    val firstname: String,
+    val lastname: String,
+    val nickname: String,
+    val email: String,
+    val phone: String,
+    val birthDate: LocalDate,
+) {
+    companion object {
+        fun fromAgent(agent: Agent): AgentDTO {
+            return AgentDTO(
+                id = agent.id!!,
+                firstname = agent.firstname,
+                lastname = agent.lastname,
+                nickname = agent.nickname,
+                email = agent.email,
+                phone = agent.phone,
+                birthDate = agent.birthDate,
+            )
+        }
+    }
+}
