+++ unRealEstate-NonBlocking-Coroutines/src/main/kotlin/dev/hosi/unrealestatenonblockingcoroutines/dtos/CreateAgentInput.kt
@@ -1,0 +1,13 @@
+package dev.hosi.unrealestatenonblockingcoroutines.dtos
+
+import java.time.LocalDate
+
+data class CreateAgentInput(
+    val firstname: String,
+    val lastname: String,
+    val nickname: String,
+    val email: String,
+    val phone: String,
+    val password: String,
+    val birthDate: LocalDate,
+)
