+++ unRealEstate-NonBlocking-Coroutines/src/main/kotlin/dev/hosi/unrealestatenonblockingcoroutines/dtos/CreatePropertyInput.kt
@@ -1,0 +1,11 @@
+package dev.hosi.unrealestatenonblockingcoroutines.dtos
+
+data class CreatePropertyInput(
+    val agentId: Long,
+    val ownerName: String,
+    val state: String,
+    val city: String,
+    val address: String,
+    val price: Long,
+    val description: String,
+)
