+++ unRealEstate-NonBlocking-Coroutines/src/test/kotlin/dev/hosi/unrealestatenonblockingcoroutines/UnRealEstateNonBlockingApplicationTests.kt
@@ -1,0 +1,12 @@
+package dev.hosi.unrealestatenonblockingcoroutines
+
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class UnRealEstateNonBlockingApplicationTests {
+
+    //    @Test
+    fun contextLoads() {
+    }
+
+}
