--- unRealEstate-NonBlocking/src/main/kotlin/dev/hosi/unreal_estate_nonblocking/UnRealEstateNonBlockingApplication.kt
@@ -1,11 +1,0 @@
-package dev.hosi.unreal_estate_nonblocking
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class UnRealEstateNonBlockingApplication
-
-fun main(args: Array<String>) {
-    runApplication<UnRealEstateNonBlockingApplication>(*args)
-}
