--- unRealEstate-NonBlocking/src/main/kotlin/dev/hosi/unreal_estate_nonblocking/dtos/CreatePropertyInput.kt
@@ -1,11 +1,0 @@
-package dev.hosi.unreal_estate_nonblocking.dtos
-
-data class CreatePropertyInput(
-    val agentId: Long,
-    val ownerName: String,
-    val state: String,
-    val city: String,
-    val address: String,
-    val price: Long,
-    val description: String,
-)
