--- unRealEstate-NonBlocking/src/test/kotlin/dev/hosi/unreal_estate_nonblocking/UnRealEstateNonBlockingApplicationTests.kt
@@ -1,13 +1,0 @@
-package dev.hosi.unreal_estate_nonblocking
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class UnRealEstateNonBlockingApplicationTests {
-
-    //    @Test
-    fun contextLoads() {
-    }
-
-}
